+++ config/phpunitgen.php
@@ -11,7 +11,7 @@ return [
      | complex tests skeleton (getter/setter tests...).
      |--------------------------------------------------------------------------
      */
-    'automaticGeneration'  => true,
+    'automaticGeneration'      => true,
 
     /*
      |--------------------------------------------------------------------------
@@ -21,7 +21,7 @@ return [
      | specific contract.
      |--------------------------------------------------------------------------
      */
-    'implementations'      => DelegateTestGenerator::implementations(),
+    'implementations'          => DelegateTestGenerator::implementations(),
 
     /*
      |--------------------------------------------------------------------------
@@ -30,7 +30,7 @@ return [
      | This string will be removed from the test class namespace.
      |--------------------------------------------------------------------------
      */
-    'baseNamespace'        => 'App',
+    'baseNamespace'            => 'App',
 
     /*
      |--------------------------------------------------------------------------
@@ -39,7 +39,7 @@ return [
      | This string will be prepend to the test class namespace.
      |--------------------------------------------------------------------------
      */
-    'baseTestNamespace'    => 'Tests',
+    'baseTestNamespace'        => 'Tests',
 
     /*
      |--------------------------------------------------------------------------
@@ -48,7 +48,7 @@ return [
      | The absolute class name to TestCase.
      |--------------------------------------------------------------------------
      */
-    'testCase'             => 'Tests\\TestCase',
+    'testCase'                 => 'Tests\\TestCase',
 
     /*
      |--------------------------------------------------------------------------
@@ -57,7 +57,7 @@ return [
      | Tells if the test class should be final.
      |--------------------------------------------------------------------------
      */
-    'testClassFinal'       => true,
+    'testClassFinal'           => true,
 
     /*
      |--------------------------------------------------------------------------
@@ -66,7 +66,16 @@ return [
      | Tells if the test class should declare strict types.
      |--------------------------------------------------------------------------
      */
-    'testClassStrictTypes' => false,
+    'testClassStrictTypes'     => false,
+
+    /*
+     |--------------------------------------------------------------------------
+     | Test class typed properties.
+     |
+     | Tells if the test class properties should be typed or documented.
+     |--------------------------------------------------------------------------
+     */
+    'testClassTypedProperties' => false,
 
     /*
      |--------------------------------------------------------------------------
@@ -77,7 +86,7 @@ return [
      | closing "/", as they will be added automatically.
      |--------------------------------------------------------------------------
      */
-    'excludedMethods'      => [
+    'excludedMethods'          => [
         '__construct',
         '__destruct',
     ],
@@ -90,7 +99,7 @@ return [
      | to the test class documentation.
      |--------------------------------------------------------------------------
      */
-    'mergedPhpDoc'         => [
+    'mergedPhpDoc'             => [
         'author',
         'copyright',
         'license',
@@ -105,7 +114,7 @@ return [
      | added to the test class documentation.
      |--------------------------------------------------------------------------
      */
-    'phpDoc'               => [],
+    'phpDoc'                   => [],
 
     /*
      |--------------------------------------------------------------------------
@@ -115,7 +124,7 @@ return [
      | contains any other useful information for test generation.
      |--------------------------------------------------------------------------
      */
-    'options'              => [
+    'options'                  => [
         /*
          |----------------------------------------------------------------------
          | Context.